CASM-3531: Implement changes to schema version handling
Atif requested that we use version range comparisons and support use of
the tilde and caret version range comparison operators.

Also change the name of the property which products will provide to
`iuf_version` in the hopes that it resolves some of the questions about
this which arose during review of this schema.

Test Description:
Ran `make` to validate schema and generate HTML docs.