CASM-3531: Refine product name, description, and version properties
Refine the name, description, and version properties in the schema. The
version has been changed so that it no longer strictly requires that it
conform to the SemVer 2.0.0 format, but is rather a requirement with a
note that it must be SemVer 2.0.0 format if the product provides a
config management repository.

The handling of `.version` files as an alternative to the `version`
property in the IUF manifest is also documented in the schema.