CASM-3564: Add schema validation code and subcommand
Add a subcommand to the main program that allows the user to validate an
IUF Product Manifest file against the IUF Product Manifest schema.

When the program is run without any arguments, the behavior is the same
as before. That is, it will start the cray-nls service that implements
the NLS and IUF API.

Bump the version of cray-nls to 0.3.0 for this new feature. Bump patch
version of cray-nls Helm chart since its functionality doesn't really
change except for pulling in a newer version of the container image.

Test Description:
Tested by running `go build` and then ran `cray-nls validate` in the
following cases:

* Validate the valid COS IUF Product Manifest file
* Introduce multiple schema violations to the COS manifest and check
  that it fails to validate, exits non-zero, and prints informative
  error messages.
* Validate a non-existent file and ensure the error is handled
  gracefully.
* Validate a file that contains bad YAML syntax and ensure the error is
  handled gracefully.