CASM-3520: Add hooks schema
Add a preliminary definition of the schema for hooks to the IUF Product
Manifest schema file.

This was developed based on discussions with Atif, John, Dean, and
others.

This probably needs further discussion and perhaps some feedback from
product teams.

An example is also provided in an `examples` directory in this repo.

Test Description:
Built with `go build`. Then ran:

```
./cray-nls validate examples/iuf-product-manifest/hooks.yaml
```

And verified it validated against the schema.