CASM-3506: Update IUF behavior for images and recipes
This change amends the descriptions for recipe and image names
to mention how the IUF will behave in the event that a recipe or image
name is duplicated in the IUF product manifest and/or content directory
manifests.

Test Description:
None technically needed, though test `cray-nls validate` against a
minimal manifest ensure the manifest schema parses.