CASM-3680: Add `/iuf/v1/stages` API endpoint
Add the `/iuf/v1/stages` API endpoint which just returns the contents of
the `stages.yaml` file being used by the IUF API. The IUF CLI wants to
use this file to dynamically determine the stages known to the IUF
operator.

A couple of whitespace changes were introduced to some of the
annotations on the API functions as a result of running the code
generation script, `scripts/code.gen.sh`.

The existing `GetStages` method of the IufController was made public so
it could be called in the API, and it was moved to a separate file.

Test Description:
Ran a local copy of cray-nls using `scripts/runDev.sh` and used Swagger
web UI to test a GET on `/iuf/v1/stages` and confirmed it responded
with the stages.yaml data.