CASM-3592: Convert workflow arguments to template input parameters
When using `templateRef` to refer to this template, it is not possible
to set the values of workflow arguments. These must be template input
parameters.

Test Description: I modified the update-product-catalog-template
workflow template on frigg to match this file, and ran the IUF CLI
to run the ims-upload operation which referenced this workflow template.