CASM-3722: Remove 'insert_component_versions' from schema
Prior to this commit, there was a boolean flag that IUF manifests could
specify to control whether or not component-level artifact data (e.g.
container images, nexus repos, etc.) was inserted into the product
catalog. The intention of this flag was for only certain products that
required backwards compatibility with prodmgr to use it.

This commit removes this flag, because the current plan is to ignore it
anyway, and for all products to insert this component-level data.

This commit increments the following versions:
* IUF manifest schema version
* cray-nls container image version
* cray-nls Helm chart version

Test Description: I validated the most recent COS IUF manifest file
against the new schema to make sure it still validates.