CASMINST-5693: Improve handling of CA certs in sat-general-template
The usage of `subPath` in a `volumeMount` causes Kubernetes to fail to
clean up old pods. This improves the handling of getting the K8s
cluster-specific certificates into the CA certs file used in Alpine
Linux without using the `subPath` in a volume.

The handling here is the same as in the `sat-podman` wrapper script. It
mounts the `platform-ca-certs.crt` file into a location read from by
`update-ca-certificates` and then it runs that command to update the
certs stored in `/etc/ssl/certs/ca-certificates.crt`, which preserves
the system-provided certs too. Then we just set `REQUESTS_CA_BUNDLE` in
the environment to ensure it is used by the Python `requests` library
used by `sat`.

Test Description:
Will be tested on frigg.